LayoutTests/inspector-obsolete/http-tests/search/search-in-static-expected.txt

Tests static content provider search.

Bug 70354  
http://127.0.0.1:8000/inspector/search/resources/search.js
Search matches: 
lineNumber: 0, line: 'function searchTestUniqueString()'
lineNumber: 3, line: '    // searchTestUniqueString two occurences on the same line searchTestUniqueString'
lineNumber: 9, line: '    searchTestUniqueString();'

Search matches: 
lineNumber: 0, line: 'function searchTestUniqueString()'
lineNumber: 3, line: '    // searchTestUniqueString two occurences on the same line searchTestUniqueString'
lineNumber: 9, line: '    searchTestUniqueString();'

Search matches: 
lineNumber: 0, line: 'function searchTestUniqueString()'
lineNumber: 3, line: '    // searchTestUniqueString two occurences on the same line searchTestUniqueString'
lineNumber: 9, line: '    searchTestUniqueString();'

Search matches: 
lineNumber: 0, line: 'function searchTestUniqueString()'
lineNumber: 3, line: '    // searchTestUniqueString two occurences on the same line searchTestUniqueString'
lineNumber: 9, line: '    searchTestUniqueString();'
